refactor: split devices-util up to match new util structure
